update directory structure
