+++ python/paddle/fluid/tests/unittests/ir/inference/CMakeLists.txt
@@ -1,76 +1,28 @@
-file(GLOB TEST_INFERENCE_IR_PASSES RELATIVE "${CMAKE_CURRENT_SOURCE_DIR}" "test_*.py")
-string(REPLACE ".py" "" TEST_INFERENCE_IR_PASSES "${TEST_INFERENCE_IR_PASSES}")
+if (WITH_GPU AND TENSORRT_FOUND)
+  file(GLOB TEST_INFERENCE_IR_PASSES RELATIVE "${CMAKE_CURRENT_SOURCE_DIR}" "test_*.py")
 
-file(GLOB TEST_TRT_IR_PASSES RELATIVE "${CMAKE_CURRENT_SOURCE_DIR}" "test_trt_*.py")
-string(REPLACE ".py" "" TEST_TRT_IR_PASSES "${TEST_TRT_IR_PASSES}")
-
-file(GLOB TEST_TRT_CONVERTER RELATIVE "${CMAKE_CURRENT_SOURCE_DIR}" "test_trt_convert_*.py")
-string(REPLACE ".py" "" TEST_TRT_CONVERTER "${TEST_TRT_CONVERTER}")
-
-foreach(TEST_INFERENCE_IR_PASS ${TEST_TRT_IR_PASSES})
-  list(REMOVE_ITEM TEST_INFERENCE_IR_PASSES ${TEST_INFERENCE_IR_PASS})
-endforeach()
-
-if(WITH_GPU AND TENSORRT_FOUND)
-  list(REMOVE_ITEM TEST_TRT_IR_PASSES test_trt_multiclass_nms_op)
-
-  foreach(TRT_CONVERT ${TEST_TRT_CONVERTER})
-    list(REMOVE_ITEM TEST_TRT_IR_PASSES ${TRT_CONVERT})
+  file(GLOB TEST_MKLDNN_OPS RELATIVE "${CMAKE_CURRENT_SOURCE_DIR}" "test_mkldnn_*.py")
+  foreach(TEST_MKLDNN_OP ${TEST_MKLDNN_OPS})
+    list(REMOVE_ITEM TEST_INFERENCE_IR_PASSES ${TEST_MKLDNN_OP})
   endforeach()
-
-  foreach(target ${TEST_TRT_IR_PASSES})
-    if(${target} STREQUAL "test_trt_slice_dynamic_plugin")
-      if("${TENSORRT_MAJOR_VERSION}.${TENSORRT_MINOR_VERSION}" VERSION_GREATER "7.1")
-        py_test_modules(${target} MODULES ${target})
-        set_tests_properties(${target} PROPERTIES TIMEOUT 60)
-      endif()
-    else()
+  if(WITH_MKLDNN)
+    foreach(target ${TEST_MKLDNN_OPS})
       py_test_modules(${target} MODULES ${target})
-    endif()
-  endforeach()
+      set_tests_properties(${target} PROPERTIES TIMEOUT 300)
+    endforeach()
+  endif()
 
-  foreach(target ${TEST_TRT_CONVERTER})
+  string(REPLACE ".py" "" TEST_INFERENCE_IR_PASSES "${TEST_INFERENCE_IR_PASSES}")
+  foreach(target ${TEST_INFERENCE_IR_PASSES})
     py_test_modules(${target} MODULES ${target})
     set_tests_properties(${target} PROPERTIES TIMEOUT 300)
   endforeach()
-endif()
 
-file(GLOB TEST_MKLDNN_IR_PASSES RELATIVE "${CMAKE_CURRENT_SOURCE_DIR}" "test_mkldnn_*.py")
-string(REPLACE ".py" "" TEST_MKLDNN_IR_PASSES "${TEST_MKLDNN_IR_PASSES}")
-foreach(TEST_INFERENCE_IR_PASS ${TEST_MKLDNN_IR_PASSES})
-  list(REMOVE_ITEM TEST_INFERENCE_IR_PASSES ${TEST_INFERENCE_IR_PASS})
-endforeach()
-
-if(WITH_MKLDNN)
-  foreach(target ${TEST_MKLDNN_IR_PASSES})
-    py_test_modules(${target} MODULES ${target})
-  endforeach()
+  if(WITH_NV_JETSON)
+    set_tests_properties(test_trt_pool_op PROPERTIES ENVIRONMENT FLAGS_fraction_of_gpu_memory_to_use=0.1 TIMEOUT 450)
+    set_tests_properties(test_trt_pool3d_op PROPERTIES ENVIRONMENT FLAGS_fraction_of_gpu_memory_to_use=0.1 TIMEOUT 450)
+  else()
+    set_tests_properties(test_trt_pool_op PROPERTIES ENVIRONMENT FLAGS_fraction_of_gpu_memory_to_use=0.1 TIMEOUT 45)
+    set_tests_properties(test_trt_pool3d_op PROPERTIES ENVIRONMENT FLAGS_fraction_of_gpu_memory_to_use=0.1 TIMEOUT 45)
+  endif()
 endif()
-
-foreach(target ${TEST_INFERENCE_IR_PASSES})
-  py_test_modules(${target} MODULES ${target})
-endforeach()
-if(WITH_GPU AND TENSORRT_FOUND)
-set_tests_properties(test_trt_subgraph_pass PROPERTIES TIMEOUT 120)
-set_tests_properties(test_trt_activation_pass PROPERTIES TIMEOUT 120)
-set_tests_properties(test_trt_conv_pass PROPERTIES TIMEOUT 120)
-#set_tests_properties(test_trt_multiclass_nms_op PROPERTIES TIMEOUT 200)
-set_tests_properties(test_trt_dynamic_shape PROPERTIES TIMEOUT 120)
-if(WITH_NV_JETSON)
-  set_tests_properties(test_trt_pool_op PROPERTIES ENVIRONMENT FLAGS_fraction_of_gpu_memory_to_use=0.1 TIMEOUT 450)
-  set_tests_properties(test_trt_pool3d_op PROPERTIES ENVIRONMENT FLAGS_fraction_of_gpu_memory_to_use=0.1 TIMEOUT 450)
-else()
-  set_tests_properties(test_trt_pool_op PROPERTIES ENVIRONMENT FLAGS_fraction_of_gpu_memory_to_use=0.1 TIMEOUT 45)
-  set_tests_properties(test_trt_pool3d_op PROPERTIES ENVIRONMENT FLAGS_fraction_of_gpu_memory_to_use=0.1 TIMEOUT 45)
-endif()
-set_tests_properties(test_trt_reduce_mean_op PROPERTIES TIMEOUT 60)
-set_tests_properties(test_trt_tile_op PROPERTIES TIMEOUT 60)
-set_tests_properties(test_trt_fc_fuse_quant_dequant_pass PROPERTIES TIMEOUT 100)
-set_tests_properties(test_trt_conv_quant_dequant_pass PROPERTIES TIMEOUT 100)
-set_tests_properties(test_trt_matmul_quant_dequant PROPERTIES TIMEOUT 100)
-set_tests_properties(test_trt_conv3d_op PROPERTIES TIMEOUT 60)
-set_tests_properties(test_trt_conv3d_transpose_op PROPERTIES TIMEOUT 60)
-set_tests_properties(test_trt_nearest_interp_v2_op PROPERTIES TIMEOUT 30)
-endif()
-
-add_subdirectory(auto_scan_test)
